style(msvc): fix C4244/C4267 in RageFileDriverMemory/NoteDataUtil/LuaManager/CryptManager (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226.

- RageFileDriverMemory.cpp (5): int m_iFilePos fed size_t byte counts
  across Read/WriteInternal; a GetFileSize() returning the backing
  buffer's size_t.
- NoteDataUtil.cpp (5, SS5-adjacent NoteData*, characterization-only):
  a pointer-diff-via-intptr_t into int (unmatched-hold warning arg); a
  size_t length into a LoadFromSMNoteDataStringWithPlayer int len
  param; a set<int>::size() added into an int track-pressed count; a
  size_t subtracted from a taps-left counter; a reverse-loop
  size()-1 bound.
- LuaManager.cpp (5): the FromStack<int>/FromStack<unsigned int>
  specializations' lua_tointeger()-into-narrower-type assignments; a
  lua_objlen() into an int thread-pool index; two more
  lua_tointeger()-into-int sites (AdjustCount, lua_pushvalues's
  upvalue arg count).
- CryptManager.cpp (5): five hash_descriptor[].process/
  rsa_verify_hash_ex calls fed a size_t buffer length where
  libtomcrypt's C API expects unsigned long.

+++ src/CryptManager.cpp
@@ -43,7 +43,7 @@ static bool HashFile( RageFileBasic &f, unsigned char buf_hash[20], int iHash )
 			return false;
 		}
 
-		iRet = hash_descriptor[iHash].process( &hash, (const unsigned char *) s.data(), s.size() );
+		iRet = hash_descriptor[iHash].process( &hash, (const unsigned char *) s.data(), static_cast<unsigned long>(s.size()) );
 		ASSERT_M( iRet == CRYPT_OK, error_to_string(iRet) );
 	}
 
@@ -340,7 +340,7 @@ bool CryptManager::Verify( RageFileBasic &file, RString sSignature, RString sPub
 	HashFile( file, buf_hash, iHash );
 
 	int iMatch;
-	int iRet = rsa_verify_hash_ex( (const unsigned char *) sSignature.data(), sSignature.size(),
+	int iRet = rsa_verify_hash_ex( (const unsigned char *) sSignature.data(), static_cast<unsigned long>(sSignature.size()),
 			buf_hash, sizeof(buf_hash),
 			LTC_PKCS_1_EMSA, iHash, 0, &iMatch, &key.m_Key );
 
@@ -391,7 +391,7 @@ RString CryptManager::GetMD5ForString( RString sData )
 
 	hash_state hash;
 	hash_descriptor[iHash].init( &hash );
-	hash_descriptor[iHash].process( &hash, (const unsigned char *) sData.data(), sData.size() );
+	hash_descriptor[iHash].process( &hash, (const unsigned char *) sData.data(), static_cast<unsigned long>(sData.size()) );
 	hash_descriptor[iHash].done( &hash, digest );
 
 	return RString( (const char *) digest, sizeof(digest) );
@@ -405,7 +405,7 @@ RString CryptManager::GetSHA1ForString( RString sData )
 
 	hash_state hash;
 	hash_descriptor[iHash].init( &hash );
-	hash_descriptor[iHash].process( &hash, (const unsigned char *) sData.data(), sData.size() );
+	hash_descriptor[iHash].process( &hash, (const unsigned char *) sData.data(), static_cast<unsigned long>(sData.size()) );
 	hash_descriptor[iHash].done( &hash, digest );
 
 	return RString( (const char *) digest, sizeof(digest) );
@@ -436,7 +436,7 @@ RString CryptManager::GetSHA256ForString( RString sData )
 
 	hash_state hash;
 	hash_descriptor[iHash].init( &hash );
-	hash_descriptor[iHash].process( &hash, (const unsigned char *) sData.data(), sData.size() );
+	hash_descriptor[iHash].process( &hash, (const unsigned char *) sData.data(), static_cast<unsigned long>(sData.size()) );
 	hash_descriptor[iHash].done( &hash, digest );
 
 	return RString( (const char *) digest, sizeof(digest) );
